# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sms Api Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api/v1"
        }
    ],
    "paths": {
        "/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login User get token",
                "operationId": "99d66635c4992aeaa6aa44ff653d0563",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "User's email",
                                        "type": "string",
                                        "example": "sinan@gmail.com"
                                    },
                                    "password": {
                                        "description": "User's password",
                                        "type": "string",
                                        "example": "1233"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LoginResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register User",
                "operationId": "08136088c4862c313b35d6518bab8d3a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "customer_id",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "customer_id": {
                                        "description": "User's customer_id",
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "name": {
                                        "description": "User's name",
                                        "type": "string",
                                        "example": "Sinan"
                                    },
                                    "email": {
                                        "description": "User's email",
                                        "type": "string",
                                        "example": "sinan@gmail.com"
                                    },
                                    "password": {
                                        "description": "User's password",
                                        "type": "string",
                                        "example": "1233"
                                    },
                                    "password_confirmation": {
                                        "description": "User's password_confirmation",
                                        "type": "string",
                                        "example": "1233"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Register success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/user": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Get User",
                "operationId": "32d03117b496ae02c2543b61438d24ad",
                "responses": {
                    "201": {
                        "description": "Get user success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResource"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnAuthorized"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "UnAuthorized": {
                "title": "UnAuthorized",
                "properties": {
                    "message": {
                        "description": "UnAuthorized Message",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "ApiResponse": {
                "title": "ApiResponse",
                "properties": {
                    "status": {
                        "description": "Status",
                        "type": "boolean"
                    },
                    "message": {
                        "description": "Message",
                        "type": "string"
                    },
                    "data": {
                        "description": "Data",
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "CustomerResource": {
                "title": "CustomerResource",
                "properties": {
                    "id": {
                        "description": "Customer Id",
                        "type": "integer"
                    },
                    "identity": {
                        "description": "Customer Identity",
                        "type": "string"
                    },
                    "name": {
                        "description": "Customer Name",
                        "type": "string"
                    },
                    "email": {
                        "description": "Customer Email",
                        "type": "string"
                    },
                    "phone": {
                        "description": "Customer Phone",
                        "type": "string"
                    },
                    "tax_no": {
                        "description": "Customer Tax No",
                        "type": "string"
                    },
                    "tax_office": {
                        "description": "Customer Tax Office",
                        "type": "string"
                    },
                    "address": {
                        "description": "Customer Address",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "LoginResource": {
                "title": "LoginResource",
                "properties": {
                    "token": {
                        "description": "Token",
                        "type": "string"
                    },
                    "expires_in": {
                        "description": "Token Expires In",
                        "type": "integer"
                    },
                    "token_type": {
                        "description": "Token Type",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "UserResource": {
                "title": "UserResource",
                "properties": {
                    "name": {
                        "description": "User Name",
                        "type": "string"
                    },
                    "email": {
                        "description": "User Email",
                        "type": "string"
                    },
                    "customer": {
                        "$ref": "#/components/schemas/CustomerResource"
                    }
                },
                "type": "object"
            }
        },
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}